Migrate save claim tests to use DB
