+++ src/test/java/uk/gov/hmcts/cmc/claimstore/controllers/utils/sampledata/SampleClaimData.java
@@ -213,5 +213,4 @@ public class SampleClaimData {
             .withAmount(SampleAmountBreakdown.validDefaults())
             .build();
     }
-
 }
